safari-600.7-branch/LayoutTests/js/dom/dfg-prototype-chain-caching-with-impure-get-own-property-slot-traps-3.html

<!DOCTYPE html>
<html>
<body>
<script src="../../resources/js-test-pre.js"></script>
<script>

description("Tests what happens when you make prototype chain accesses with impure GetOwnPropertySlot traps in the way.");

var img = new Image();
img.name = "getElementsByTagName";
document.body.appendChild(img);

function f() {
    return document.getElementsByTagName;
}

dfgShouldBe(f, "typeof f()", "\"object\"");
shouldBe("document.body.removeChild(img); typeof f()", "\"function\"")

</script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
