Safari-611.1.10/LayoutTests/js/dom/dfg-prototype-chain-caching-with-impure-get-own-property-slot-traps-3-expected.txt

Tests what happens when you make prototype chain accesses with impure GetOwnPropertySlot traps in the way.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS typeof f() is "object" on all iterations including after DFG tier-up.
PASS document.body.removeChild(img); typeof f() is "function"
PASS successfullyParsed is true

TEST COMPLETE
